+++ fab_provision/provision.py
@@ -9,7 +9,7 @@ def provision():
         install_git()
     install_chef()
     put_config()
-    run_berkshelf()
     put_cookbooks()
+    run_berkshelf()
     put_data_bags()
     run_chef()