+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Added new config `chef_dk` - set ChefDK version or `current`
 - Added config `install_git`, e.g. for installing cookbooks from git
 
+### Fixed
+- Fixed dependency issues by sourcing cookbooks from local directory
+
 ## [0.0.2] - 2016-05-20
 ### Fixed
 - Fix bug if ChefDK is not installed